extract common agg interfaces
