+++ modules/calcite/src/test/java/org/apache/ignite/internal/processors/query/calcite/planner/SortAggregatePlannerTest.java
@@ -17,10 +17,24 @@
 
 package org.apache.ignite.internal.processors.query.calcite.planner;
 
+import java.util.Arrays;
+
 import org.apache.calcite.plan.RelOptPlanner;
+import org.apache.calcite.plan.RelOptUtil;
 import org.apache.calcite.rel.RelCollations;
+import org.apache.calcite.rel.type.RelDataTypeFactory;
 import org.apache.calcite.util.ImmutableIntList;
+import org.apache.ignite.internal.processors.query.calcite.metadata.ColocationGroup;
+import org.apache.ignite.internal.processors.query.calcite.prepare.PlanningContext;
+import org.apache.ignite.internal.processors.query.calcite.rel.IgniteRel;
+import org.apache.ignite.internal.processors.query.calcite.rel.IgniteSort;
+import org.apache.ignite.internal.processors.query.calcite.rel.agg.IgniteReduceSortAggregate;
+import org.apache.ignite.internal.processors.query.calcite.rel.agg.IgniteSingleSortAggregate;
 import org.apache.ignite.internal.processors.query.calcite.schema.IgniteSchema;
+import org.apache.ignite.internal.processors.query.calcite.trait.IgniteDistribution;
+import org.apache.ignite.internal.processors.query.calcite.trait.IgniteDistributions;
+import org.apache.ignite.internal.processors.query.calcite.type.IgniteTypeFactory;
+import org.apache.ignite.internal.processors.query.calcite.type.IgniteTypeSystem;
 import org.apache.ignite.testframework.GridTestUtils;
 import org.junit.Test;
 
@@ -46,10 +60,108 @@ public class SortAggregatePlannerTest extends AbstractAggregatePlannerTest {
             () -> physicalPlan(
                 sqlMin,
                 publicSchema,
-                "HashAggregateConverterRule"
+                "HashSingleAggregateConverterRule", "HashMapReduceAggregateConverterRule"
             ),
             RelOptPlanner.CannotPlanException.class,
             "There are not enough rules to produce a node with desired properties"
+        );
+    }
+
+    /**
+     * @throws Exception If failed.
+     */
+    @Test
+    public void collationPermuteSingle() throws Exception {
+        IgniteTypeFactory f = new IgniteTypeFactory(IgniteTypeSystem.INSTANCE);
+
+        TestTable tbl = new TestTable(
+            new RelDataTypeFactory.Builder(f)
+                .add("ID", f.createJavaType(Integer.class))
+                .add("VAL0", f.createJavaType(Integer.class))
+                .add("VAL1", f.createJavaType(Integer.class))
+                .add("GRP0", f.createJavaType(Integer.class))
+                .add("GRP1", f.createJavaType(Integer.class))
+                .build()) {
+
+            @Override public IgniteDistribution distribution() {
+                return IgniteDistributions.broadcast();
+            }
+        }
+            .addIndex(RelCollations.of(ImmutableIntList.of(3, 4)), "grp0_1");
+
+        IgniteSchema publicSchema = new IgniteSchema("PUBLIC");
+
+        publicSchema.addTable("TEST", tbl);
+
+        String sql = "SELECT MIN(val0) FROM test GROUP BY grp1, grp0";
+
+        IgniteRel phys = physicalPlan(
+            sql,
+            publicSchema,
+            "HashSingleAggregateConverterRule", "HashMapReduceAggregateConverterRule"
+        );
+
+        IgniteSingleSortAggregate agg = findFirstNode(phys, byClass(IgniteSingleSortAggregate.class));
+
+        assertNotNull("Invalid plan\n" + RelOptUtil.toString(phys), agg);
+
+        assertNull(
+            "Invalid plan\n" + RelOptUtil.toString(phys),
+            findFirstNode(phys, byClass(IgniteSort.class))
+        );
+    }
+
+    /**
+     * @throws Exception If failed.
+     */
+    @Test
+    public void collationPermuteMapReduce() throws Exception {
+        IgniteTypeFactory f = new IgniteTypeFactory(IgniteTypeSystem.INSTANCE);
+
+        TestTable tbl = new TestTable(
+            new RelDataTypeFactory.Builder(f)
+                .add("ID", f.createJavaType(Integer.class))
+                .add("VAL0", f.createJavaType(Integer.class))
+                .add("VAL1", f.createJavaType(Integer.class))
+                .add("GRP0", f.createJavaType(Integer.class))
+                .add("GRP1", f.createJavaType(Integer.class))
+                .build()) {
+
+            @Override public ColocationGroup colocationGroup(PlanningContext ctx) {
+                return ColocationGroup.forAssignments(Arrays.asList(
+                    select(nodes, 0, 1),
+                    select(nodes, 1, 2),
+                    select(nodes, 2, 0),
+                    select(nodes, 0, 1),
+                    select(nodes, 1, 2)
+                ));
+            }
+
+            @Override public IgniteDistribution distribution() {
+                return IgniteDistributions.affinity(0, "test", "hash");
+            }
+        }
+            .addIndex(RelCollations.of(ImmutableIntList.of(3, 4)), "grp0_1");
+
+        IgniteSchema publicSchema = new IgniteSchema("PUBLIC");
+
+        publicSchema.addTable("TEST", tbl);
+
+        String sql = "SELECT MIN(val0) FROM test GROUP BY grp1, grp0";
+
+        IgniteRel phys = physicalPlan(
+            sql,
+            publicSchema,
+            "HashSingleAggregateConverterRule", "HashMapReduceAggregateConverterRule"
+        );
+
+        IgniteReduceSortAggregate agg = findFirstNode(phys, byClass(IgniteReduceSortAggregate.class));
+
+        assertNotNull("Invalid plan\n" + RelOptUtil.toString(phys), agg);
+
+        assertNull(
+            "Invalid plan\n" + RelOptUtil.toString(phys),
+            findFirstNode(phys, byClass(IgniteSort.class))
         );
     }
 }
